pyproject.toml
[project]
name = "email-assistant"
version = "0.1.0"
description = "LangGraph-based email assistant with triage, HITL, memory, and Gmail integration."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langgraph-checkpoint-postgres",
    "langchain-openai",
    "langchain-core",
    "supabase",
    "psycopg[binary,pool]",
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "langgraph-cli[inmem]",
    "langgraph-api>=0.7.59",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/email_assistant"]

